fix: address github account review follow-ups
