+++ api/app/src/__tests__/github-user-account-refresh.test.ts
@@ -38,10 +38,8 @@ vi.mock("@repo/github-app-node", () => ({
 }));
 
 vi.mock("../env", () => ({
-  env: {
-    ENCRYPTION_KEY:
-      "0000000000000000000000000000000000000000000000000000000000000000",
-  },
+  getAppEncryptionKey: () =>
+    "0000000000000000000000000000000000000000000000000000000000000000",
 }));
 
 vi.mock("../services/github/config", () => ({
