+++ api/app/src/services/github/user-account/finalize-account.ts
@@ -2,7 +2,7 @@ import { finalizeActiveUserSourceControlAccount } from "@db/app";
 import { db } from "@db/app/client";
 import { encrypt } from "@repo/app-encryption";
 
-import { env } from "../../../env";
+import { getAppEncryptionKey } from "../../../env";
 
 export async function finalizeGitHubUserAccountBinding(input: {
   accessToken: string;
@@ -12,9 +12,10 @@ export async function finalizeGitHubUserAccountBinding(input: {
   refreshToken: string;
   refreshTokenExpiresAt: Date;
 }) {
+  const encryptionKey = getAppEncryptionKey();
   const [encryptedAccessToken, encryptedRefreshToken] = await Promise.all([
-    encrypt(input.accessToken, env.ENCRYPTION_KEY),
-    encrypt(input.refreshToken, env.ENCRYPTION_KEY),
+    encrypt(input.accessToken, encryptionKey),
+    encrypt(input.refreshToken, encryptionKey),
   ]);
 
   return finalizeActiveUserSourceControlAccount(db, {
