+++ api/app/src/services/github/user-account/refresh.ts
@@ -11,7 +11,7 @@ import {
   refreshGitHubUserAccessToken,
 } from "@repo/github-app-node";
 
-import { env } from "../../../env";
+import { getAppEncryptionKey } from "../../../env";
 import { getGitHubAppConfig } from "../config";
 
 const DEFAULT_REFRESH_WINDOW_MS = 60 * 60 * 1000;
@@ -49,15 +49,13 @@ export async function getFreshGitHubUserAccessToken(input: {
 
   const now = input.now?.() ?? new Date();
   const refreshWindowMs = input.refreshWindowMs ?? DEFAULT_REFRESH_WINDOW_MS;
+  const encryptionKey = getAppEncryptionKey();
   if (
     account.accessTokenExpiresAt.getTime() - now.getTime() >
     refreshWindowMs
   ) {
     return {
-      accessToken: await decrypt(
-        account.encryptedAccessToken,
-        env.ENCRYPTION_KEY
-      ),
+      accessToken: await decrypt(account.encryptedAccessToken, encryptionKey),
     };
   }
 
@@ -73,7 +71,7 @@ export async function getFreshGitHubUserAccessToken(input: {
 
   const refreshToken = await decrypt(
     account.encryptedRefreshToken,
-    env.ENCRYPTION_KEY
+    encryptionKey
   );
 
   let refreshed: Awaited<ReturnType<typeof refreshGitHubUserAccessToken>>;
@@ -92,6 +90,7 @@ export async function getFreshGitHubUserAccessToken(input: {
       now,
       observedEncryptedRefreshToken: account.encryptedRefreshToken,
       refreshWindowMs,
+      encryptionKey,
     });
     if (recovered) {
       return recovered;
@@ -122,8 +121,8 @@ export async function getFreshGitHubUserAccessToken(input: {
     now.getTime() + refreshed.refreshTokenExpiresIn * 1000
   );
   const [encryptedAccessToken, encryptedRefreshToken] = await Promise.all([
-    encrypt(refreshed.accessToken, env.ENCRYPTION_KEY),
-    encrypt(refreshed.refreshToken, env.ENCRYPTION_KEY),
+    encrypt(refreshed.accessToken, encryptionKey),
+    encrypt(refreshed.refreshToken, encryptionKey),
   ]);
 
   const updated = await updateObservedUserSourceControlAccountTokens(input.db, {
@@ -144,6 +143,7 @@ export async function getFreshGitHubUserAccessToken(input: {
       now,
       observedEncryptedRefreshToken: account.encryptedRefreshToken,
       refreshWindowMs,
+      encryptionKey,
     });
     if (recovered) {
       return recovered;
@@ -157,6 +157,7 @@ export async function getFreshGitHubUserAccessToken(input: {
 async function recoverFromConcurrentRefresh(input: {
   clerkUserId: string;
   db: Database;
+  encryptionKey: string;
   now: Date;
   observedEncryptedRefreshToken: string;
   refreshWindowMs: number;
@@ -178,7 +179,7 @@ async function recoverFromConcurrentRefresh(input: {
   return {
     accessToken: await decrypt(
       currentAccount.encryptedAccessToken,
-      env.ENCRYPTION_KEY
+      input.encryptionKey
     ),
   };
 }
