+++ apps/app/turbo.json
@@ -26,7 +26,6 @@
         "CLERK_SECRET_KEY",
         "CLERK_CLI_OAUTH_CLIENT_ID",
         "CLERK_DESKTOP_OAUTH_CLIENT_ID",
-        "ENCRYPTION_KEY",
         "DATABASE_HOST",
         "DATABASE_USERNAME",
         "DATABASE_PASSWORD",
