+++ emulators/github/src/__tests__/server.test.ts
@@ -1,4 +1,4 @@
-import { createHash, createHmac, createPrivateKey } from "node:crypto";
+import { createHmac } from "node:crypto";
 import { createServer } from "node:http";
 import { Store } from "@emulators/core";
 import {
@@ -6,8 +6,7 @@ import {
   githubPlugin,
   seedFromConfig,
 } from "@emulators/github";
-import { SignJWT } from "jose";
-import { afterAll, beforeAll, describe, expect, it, vi } from "vitest";
+import { afterAll, beforeAll, describe, expect, it } from "vitest";
 
 import {
   createGitHubEmulatorSeed,
@@ -19,145 +18,20 @@ import { createGitHubCompatibleFetch } from "../github-compatible-routes";
 import {
   addOrgMembership,
   type StartedGitHubEmulator,
-  type StartGitHubEmulatorInput,
   startGitHubEmulator,
 } from "../server";
+import {
+  appCallbackUrl,
+  authorizeOAuthCode,
+  createAppJwt,
+  createCodeChallenge,
+  exchangeOAuthCode,
+  mintOAuthToken,
+  startGitHubEmulatorOnAvailablePort,
+} from "./test-helpers";
 
 let emulator: StartedGitHubEmulator | undefined;
 let emulatorPort: number;
-
-const TEST_PORT_MIN = 40_000;
-const TEST_PORT_SPAN = 10_000;
-const TEST_PORT_ATTEMPTS = 20;
-
-function getRandomTestPort() {
-  return TEST_PORT_MIN + Math.floor(Math.random() * TEST_PORT_SPAN);
-}
-
-function isAddrInUse(error: unknown): error is NodeJS.ErrnoException {
-  return (
-    typeof error === "object" &&
-    error !== null &&
-    "code" in error &&
-    (error as NodeJS.ErrnoException).code === "EADDRINUSE"
-  );
-}
-
-async function startGitHubEmulatorOnAvailablePort(
-  input: Omit<StartGitHubEmulatorInput, "port"> = {}
-) {
-  let lastAddrInUseError: NodeJS.ErrnoException | undefined;
-
-  for (let attempt = 0; attempt < TEST_PORT_ATTEMPTS; attempt += 1) {
-    try {
-      return await startGitHubEmulator({
-        ...input,
-        port: getRandomTestPort(),
-      });
-    } catch (error) {
-      if (!isAddrInUse(error)) {
-        throw error;
-      }
-      lastAddrInUseError = error;
-    }
-  }
-
-  throw (
-    lastAddrInUseError ??
-    new Error("Failed to start GitHub emulator on an available local port")
-  );
-}
-
-async function createAppJwt() {
-  const key = createPrivateKey(GITHUB_EMULATOR_FIXTURES.githubAppPrivateKey);
-  const now = Math.floor(Date.now() / 1000);
-  return await new SignJWT({})
-    .setProtectedHeader({ alg: "RS256" })
-    .setIssuedAt(now - 30)
-    .setExpirationTime(now + 9 * 60)
-    .setIssuer(String(GITHUB_EMULATOR_FIXTURES.githubAppId))
-    .sign(key);
-}
-
-function createCodeChallenge(verifier: string) {
-  return createHash("sha256").update(verifier).digest("base64url");
-}
-
-function appCallbackUrl(path = "/api/github/oauth/callback") {
-  return new URL(path, "https://lightfast.localhost").toString();
-}
-
-function userAccountCallbackUrl() {
-  return appCallbackUrl("/api/github/user/oauth/callback");
-}
-
-async function authorizeOAuthCode(
-  codeVerifier: string,
-  input: {
-    redirectUri?: string;
-    state?: string;
-  } = {}
-) {
-  const redirectUri = input.redirectUri ?? appCallbackUrl();
-  const state = input.state ?? "oauth_state_123";
-  const authorizeUrl = new URL(`${emulator?.url}/login/oauth/authorize`);
-  authorizeUrl.searchParams.set(
-    "client_id",
-    GITHUB_EMULATOR_FIXTURES.oauthClientId
-  );
-  authorizeUrl.searchParams.set("redirect_uri", redirectUri);
-  authorizeUrl.searchParams.set("state", state);
-  authorizeUrl.searchParams.set(
-    "code_challenge",
-    createCodeChallenge(codeVerifier)
-  );
-  authorizeUrl.searchParams.set("code_challenge_method", "S256");
-
-  const authorizeRes = await fetch(authorizeUrl, { redirect: "manual" });
-  expect(authorizeRes.status).toBe(302);
-  const callback = new URL(authorizeRes.headers.get("location") ?? "");
-  expect(callback.origin + callback.pathname).toBe(redirectUri);
-  expect(callback.searchParams.get("state")).toBe(state);
-  const code = callback.searchParams.get("code");
-  expect(code).toEqual(expect.any(String));
-  return code ?? "";
-}
-
-async function exchangeOAuthCode(
-  code: string,
-  codeVerifier: string,
-  input: {
-    redirectUri?: string;
-  } = {}
-) {
-  return await fetch(`${emulator?.url}/login/oauth/access_token`, {
-    method: "POST",
-    headers: {
-      accept: "application/json",
-      "content-type": "application/json",
-    },
-    body: JSON.stringify({
-      client_id: GITHUB_EMULATOR_FIXTURES.oauthClientId,
-      client_secret: GITHUB_EMULATOR_FIXTURES.oauthClientSecret,
-      code,
-      code_verifier: codeVerifier,
-      redirect_uri: input.redirectUri ?? appCallbackUrl(),
-    }),
-  });
-}
-
-async function mintOAuthToken() {
-  const codeVerifier = "verifier_123456789012345678901234567890";
-  const code = await authorizeOAuthCode(codeVerifier);
-  const tokenRes = await exchangeOAuthCode(code, codeVerifier);
-  expect(tokenRes.status).toBe(200);
-  const body = (await tokenRes.json()) as { access_token?: string };
-  expect(body).toMatchObject({
-    access_token: expect.stringMatching(/^gho_/),
-    token_type: "bearer",
-  });
-  return body.access_token ?? "";
-}
 
 beforeAll(async () => {
   emulator = await startGitHubEmulatorOnAvailablePort();
@@ -383,8 +257,12 @@ describe("@repo/github-emulator", () => {
 
   it("performs OAuth authorize and token exchange with PKCE", async () => {
     const codeVerifier = "verifier_123456789012345678901234567890";
-    const code = await authorizeOAuthCode(codeVerifier);
-    const tokenRes = await exchangeOAuthCode(code, codeVerifier);
+    const code = await authorizeOAuthCode(emulator?.url ?? "", codeVerifier);
+    const tokenRes = await exchangeOAuthCode(
+      emulator?.url ?? "",
+      code,
+      codeVerifier
+    );
 
     expect(tokenRes.status).toBe(200);
     await expect(tokenRes.json()).resolves.toMatchObject({
@@ -393,353 +271,48 @@ describe("@repo/github-emulator", () => {
     });
   });
 
-  it("performs user account OAuth with refreshable token fields", async () => {
-    const codeVerifier = "verifier_123456789012345678901234567890";
-    const code = await authorizeOAuthCode(codeVerifier, {
-      redirectUri: userAccountCallbackUrl(),
-      state: "user_account_state_123",
-    });
-    const tokenRes = await exchangeOAuthCode(code, codeVerifier, {
-      redirectUri: userAccountCallbackUrl(),
-    });
-
-    expect(tokenRes.status).toBe(200);
-    await expect(tokenRes.json()).resolves.toMatchObject({
-      access_token: expect.stringMatching(/^ghu_/),
-      expires_in: 28_800,
-      refresh_token: expect.stringMatching(/^ghr_/),
-      refresh_token_expires_in: 15_768_000,
-      token_type: "bearer",
-    });
-  });
-
-  it("refreshes user account OAuth tokens", async () => {
-    const codeVerifier = "verifier_123456789012345678901234567890";
-    const code = await authorizeOAuthCode(codeVerifier, {
-      redirectUri: userAccountCallbackUrl(),
-      state: "user_account_state_123",
-    });
-    const tokenRes = await exchangeOAuthCode(code, codeVerifier, {
-      redirectUri: userAccountCallbackUrl(),
-    });
-    expect(tokenRes.status).toBe(200);
-    const tokenBody = (await tokenRes.json()) as {
-      access_token?: string;
-      refresh_token?: string;
-    };
-    expect(tokenBody).toMatchObject({
-      access_token: expect.stringMatching(/^ghu_/),
-      refresh_token: expect.stringMatching(/^ghr_/),
-    });
-
-    const refreshRes = await fetch(
-      `${emulator?.url}/login/oauth/access_token`,
-      {
-        method: "POST",
-        headers: {
-          accept: "application/json",
-          "content-type": "application/json",
-        },
-        body: JSON.stringify({
-          client_id: GITHUB_EMULATOR_FIXTURES.oauthClientId,
-          client_secret: GITHUB_EMULATOR_FIXTURES.oauthClientSecret,
-          grant_type: "refresh_token",
-          refresh_token: tokenBody.refresh_token,
-        }),
-      }
-    );
-
-    expect(refreshRes.status).toBe(200);
-    const refreshBody = (await refreshRes.json()) as {
-      access_token?: string;
-      refresh_token?: string;
-    };
-    expect(refreshBody).toMatchObject({
-      access_token: expect.stringMatching(/^ghu_/),
-      expires_in: 28_800,
-      refresh_token: expect.stringMatching(/^ghr_/),
-      refresh_token_expires_in: 15_768_000,
-      token_type: "bearer",
-    });
-    expect(refreshBody.access_token).not.toBe(tokenBody.access_token);
-    expect(refreshBody.refresh_token).not.toBe(tokenBody.refresh_token);
-
-    const userRes = await fetch(`${emulator?.url}/user`, {
-      headers: {
-        authorization: `Bearer ${refreshBody.access_token}`,
-      },
-    });
-    expect(userRes.status).toBe(200);
-    await expect(userRes.json()).resolves.toMatchObject({
-      login: GITHUB_EMULATOR_FIXTURES.githubUserLogin,
-      type: "User",
-    });
-  });
-
-  it("revokes user account OAuth grants", async () => {
-    const codeVerifier = "verifier_123456789012345678901234567890";
-    const code = await authorizeOAuthCode(codeVerifier, {
-      redirectUri: userAccountCallbackUrl(),
-      state: "user_account_state_123",
-    });
-    const tokenRes = await exchangeOAuthCode(code, codeVerifier, {
-      redirectUri: userAccountCallbackUrl(),
-    });
-    expect(tokenRes.status).toBe(200);
-    const tokenBody = (await tokenRes.json()) as {
-      access_token?: string;
-      refresh_token?: string;
-    };
-
-    const revokeRes = await fetch(
-      `${emulator?.url}/applications/${GITHUB_EMULATOR_FIXTURES.oauthClientId}/grant`,
-      {
-        method: "DELETE",
-        headers: {
-          accept: "application/vnd.github+json",
-          authorization: `Basic ${Buffer.from(
-            `${GITHUB_EMULATOR_FIXTURES.oauthClientId}:${GITHUB_EMULATOR_FIXTURES.oauthClientSecret}`
-          ).toString("base64")}`,
-          "content-type": "application/json",
-        },
-        body: JSON.stringify({ access_token: tokenBody.access_token }),
-      }
-    );
-    expect(revokeRes.status).toBe(204);
-
-    const userRes = await fetch(`${emulator?.url}/user`, {
-      headers: {
-        authorization: `Bearer ${tokenBody.access_token}`,
-      },
-    });
-    expect(userRes.status).toBe(401);
-
-    const refreshRes = await fetch(
-      `${emulator?.url}/login/oauth/access_token`,
-      {
-        method: "POST",
-        headers: {
-          accept: "application/json",
-          "content-type": "application/json",
-        },
-        body: JSON.stringify({
-          client_id: GITHUB_EMULATOR_FIXTURES.oauthClientId,
-          client_secret: GITHUB_EMULATOR_FIXTURES.oauthClientSecret,
-          grant_type: "refresh_token",
-          refresh_token: tokenBody.refresh_token,
-        }),
-      }
-    );
-    expect(refreshRes.status).toBe(200);
-    await expect(refreshRes.json()).resolves.toMatchObject({
-      error: "bad_refresh_token",
-    });
-  });
-
-  it("refreshes user account OAuth tokens after the access token expires", async () => {
-    vi.useFakeTimers({ toFake: ["Date"] });
-    vi.setSystemTime(new Date("2026-05-30T00:00:00.000Z"));
-
-    try {
-      const store = new Store();
-      githubPlugin.seed?.(store, GITHUB_EMULATOR_FIXTURES.origin);
-      seedFromConfig(
-        store,
-        GITHUB_EMULATOR_FIXTURES.origin,
-        createGitHubEmulatorSeed()
-      );
-      addOrgMembership(store);
-      const tokenMap = new Map();
-      const fetchCompatible = createGitHubCompatibleFetch({
-        appOrigin: "https://lightfast.localhost",
-        fallbackFetch: () =>
-          Response.json({ message: "fallback" }, { status: 418 }),
-        publicOrigin: GITHUB_EMULATOR_FIXTURES.origin,
-        resetStore: () => undefined,
-        store,
-        tokenMap,
-      });
-
-      const codeVerifier = "verifier_123456789012345678901234567890";
-      const authorizeUrl = new URL(
-        `${GITHUB_EMULATOR_FIXTURES.origin}/login/oauth/authorize`
-      );
-      authorizeUrl.searchParams.set(
-        "client_id",
-        GITHUB_EMULATOR_FIXTURES.oauthClientId
-      );
-      authorizeUrl.searchParams.set("redirect_uri", userAccountCallbackUrl());
-      authorizeUrl.searchParams.set("state", "user_account_state_123");
-      authorizeUrl.searchParams.set(
-        "code_challenge",
-        createCodeChallenge(codeVerifier)
-      );
-      authorizeUrl.searchParams.set("code_challenge_method", "S256");
-      const authorizeRes = await fetchCompatible(new Request(authorizeUrl));
-      expect(authorizeRes.status).toBe(302);
-      const callback = new URL(authorizeRes.headers.get("location") ?? "");
-
-      const tokenRes = await fetchCompatible(
-        new Request(
-          `${GITHUB_EMULATOR_FIXTURES.origin}/login/oauth/access_token`,
-          {
-            method: "POST",
-            headers: {
-              accept: "application/json",
-              "content-type": "application/json",
-            },
-            body: JSON.stringify({
-              client_id: GITHUB_EMULATOR_FIXTURES.oauthClientId,
-              client_secret: GITHUB_EMULATOR_FIXTURES.oauthClientSecret,
-              code: callback.searchParams.get("code"),
-              code_verifier: codeVerifier,
-              redirect_uri: userAccountCallbackUrl(),
-            }),
-          }
-        )
-      );
-      expect(tokenRes.status).toBe(200);
-      const tokenBody = (await tokenRes.json()) as {
-        access_token?: string;
-        refresh_token?: string;
-      };
-      expect(tokenBody).toMatchObject({
-        access_token: expect.stringMatching(/^ghu_/),
-        refresh_token: expect.stringMatching(/^ghr_/),
-      });
-
-      vi.advanceTimersByTime(28_800 * 1000 + 1);
-
-      const userRes = await fetchCompatible(
-        new Request(`${GITHUB_EMULATOR_FIXTURES.origin}/user`, {
-          headers: {
-            authorization: `Bearer ${tokenBody.access_token}`,
-          },
-        })
-      );
-      expect(userRes.status).toBe(401);
-
-      const refreshRes = await fetchCompatible(
-        new Request(
-          `${GITHUB_EMULATOR_FIXTURES.origin}/login/oauth/access_token`,
-          {
-            method: "POST",
-            headers: {
-              accept: "application/json",
-              "content-type": "application/json",
-            },
-            body: JSON.stringify({
-              client_id: GITHUB_EMULATOR_FIXTURES.oauthClientId,
-              client_secret: GITHUB_EMULATOR_FIXTURES.oauthClientSecret,
-              grant_type: "refresh_token",
-              refresh_token: tokenBody.refresh_token,
-            }),
-          }
-        )
-      );
-      expect(refreshRes.status).toBe(200);
-      await expect(refreshRes.json()).resolves.toMatchObject({
-        access_token: expect.stringMatching(/^ghu_/),
-        expires_in: 28_800,
-        refresh_token: expect.stringMatching(/^ghr_/),
-        refresh_token_expires_in: 15_768_000,
-      });
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  it("rejects expired user account OAuth tokens on fallback routes", async () => {
-    vi.useFakeTimers({ toFake: ["Date"] });
-    vi.setSystemTime(new Date("2026-05-30T00:00:00.000Z"));
-
-    try {
-      const codeVerifier = "verifier_123456789012345678901234567890";
-      const code = await authorizeOAuthCode(codeVerifier, {
-        redirectUri: userAccountCallbackUrl(),
-        state: "user_account_state_123",
-      });
-      const tokenRes = await exchangeOAuthCode(code, codeVerifier, {
-        redirectUri: userAccountCallbackUrl(),
-      });
-      expect(tokenRes.status).toBe(200);
-      const tokenBody = (await tokenRes.json()) as {
-        access_token?: string;
-        refresh_token?: string;
-      };
-      expect(tokenBody).toMatchObject({
-        access_token: expect.stringMatching(/^ghu_/),
-        refresh_token: expect.stringMatching(/^ghr_/),
-      });
-
-      vi.setSystemTime(new Date("2026-05-30T08:00:01.000Z"));
-
-      const userRes = await fetch(`${emulator?.url}/user`, {
-        headers: {
-          authorization: `Bearer ${tokenBody.access_token}`,
-        },
-      });
-      expect(userRes.status).toBe(401);
-
-      const orgsRes = await fetch(`${emulator?.url}/user/orgs`, {
-        headers: {
-          authorization: `Bearer ${tokenBody.access_token}`,
-        },
-      });
-      expect(orgsRes.status).toBe(401);
-      await expect(orgsRes.json()).resolves.toEqual({
-        message: "Bad credentials",
-      });
-
-      const refreshRes = await fetch(
-        `${emulator?.url}/login/oauth/access_token`,
-        {
-          method: "POST",
-          headers: {
-            accept: "application/json",
-            "content-type": "application/json",
-          },
-          body: JSON.stringify({
-            client_id: GITHUB_EMULATOR_FIXTURES.oauthClientId,
-            client_secret: GITHUB_EMULATOR_FIXTURES.oauthClientSecret,
-            grant_type: "refresh_token",
-            refresh_token: tokenBody.refresh_token,
-          }),
-        }
-      );
-      expect(refreshRes.status).toBe(200);
-      await expect(refreshRes.json()).resolves.toMatchObject({
-        access_token: expect.stringMatching(/^ghu_/),
-        refresh_token: expect.stringMatching(/^ghr_/),
-      });
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
   it("rejects bad PKCE verifiers and one-time OAuth code reuse", async () => {
     const codeVerifier = "verifier_123456789012345678901234567890";
-    const code = await authorizeOAuthCode(codeVerifier);
-    const badPkceRes = await exchangeOAuthCode(code, "wrong_verifier");
+    const code = await authorizeOAuthCode(emulator?.url ?? "", codeVerifier);
+    const badPkceRes = await exchangeOAuthCode(
+      emulator?.url ?? "",
+      code,
+      "wrong_verifier"
+    );
     expect(badPkceRes.status).toBe(200);
     await expect(badPkceRes.json()).resolves.toMatchObject({
       error: "bad_verification_code",
     });
 
-    const firstGoodRes = await exchangeOAuthCode(code, codeVerifier);
+    const firstGoodRes = await exchangeOAuthCode(
+      emulator?.url ?? "",
+      code,
+      codeVerifier
+    );
     expect(firstGoodRes.status).toBe(200);
     await expect(firstGoodRes.json()).resolves.toMatchObject({
       error: "bad_verification_code",
     });
 
-    const reusableCode = await authorizeOAuthCode(codeVerifier);
-    const firstUseRes = await exchangeOAuthCode(reusableCode, codeVerifier);
+    const reusableCode = await authorizeOAuthCode(
+      emulator?.url ?? "",
+      codeVerifier
+    );
+    const firstUseRes = await exchangeOAuthCode(
+      emulator?.url ?? "",
+      reusableCode,
+      codeVerifier
+    );
     expect(firstUseRes.status).toBe(200);
     await expect(firstUseRes.json()).resolves.toMatchObject({
       access_token: expect.stringMatching(/^gho_/),
     });
 
-    const secondUseRes = await exchangeOAuthCode(reusableCode, codeVerifier);
+    const secondUseRes = await exchangeOAuthCode(
+      emulator?.url ?? "",
+      reusableCode,
+      codeVerifier
+    );
     expect(secondUseRes.status).toBe(200);
     await expect(secondUseRes.json()).resolves.toMatchObject({
       error: "bad_verification_code",
@@ -747,7 +320,7 @@ describe("@repo/github-emulator", () => {
   });
 
   it("returns the OAuth user and accessible app installations", async () => {
-    const oauthToken = await mintOAuthToken();
+    const oauthToken = await mintOAuthToken(emulator?.url ?? "");
     const userRes = await fetch(`${emulator?.url}/user`, {
       headers: {
         authorization: `Bearer ${oauthToken}`,
